+++ bin/store/src/auth/structs.rs
@@ -5,6 +5,7 @@ use serde_json::Value;
 pub struct LoginResponse {
     pub message: String,
     pub token: Option<String>,
+    pub account_organization_id: Option<String>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -102,4 +103,9 @@ pub struct Session {
     pub role_permissions: Option<SessionPermissionCache>,
     pub record_permissions: Option<SessionPermissionCache>,
     pub valid_pass_keys: Option<SessionPermissionCache>,
+    #[serde(default)]
+    pub ip_address: Option<String>,
+    #[serde(default)]
+    pub location: Option<String>,
+    pub account_profile_id: Option<String>,
 }