+++ bin/store/src/models/mod.rs
@@ -35,6 +35,7 @@ pub mod record_permission_model;
 pub mod role_permission_model;
 pub mod sample_model;
 pub mod session_model;
+pub mod signed_in_activity_model;
 pub mod stream_queue_item_model;
 pub mod stream_queue_model;
 pub mod sync_endpoint_model;